Moved all ClassicModeDeluxe resources into a single folder
Apparently map scripts can load from the plugins folder.

The models/sounds/sprites were moved in preparation for converting this to a separate repo, since it's complicated to work with multiple repos in the same folder (svencoop_addon in this case). The install.bat file will move everything into the right place.