prevent double map loading by toggling classic mode early
If a game_end is triggered, then a map series or single map must have ended. In this case, it's likely that classic mode could need to be changed. This will check if the next map in the cycle is classic and set the mode accordingly before the level change starts. This prevents the level restart that normally happens when toggling classic mode.

trigger_changelevel shouldn't do this because map series can't be both modern and classic. It's still possible that an admin uses the changelevel command or a map acts as a hub world to change change to many different types of maps. So, the overly complicated level restart logic is still here. More hacks stacked on hacks.